Refactor(Currency): move currency schema into a schema folder
